requirements.txt
Werkzeug==2.2.2
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
pandas==1.3.2
mysql-connector-python==8.0.26
python-dotenv==0.19.0
numpy==1.26.4
